+++ src/shared_utilities/perturbed_drivers.jl
@@ -224,6 +224,9 @@ function prescribed_perturbed_temperature_era5(
         start_date,
         regridder_type,
         regridder_kwargs = (; interpolation_method),
+        file_reader_kwargs = (;
+            preprocess_func = (data) -> max.(data, Float32(0)),
+        ),
         method = time_interpolation_method,
     )
     function compute_diffuse_fraction(total, direct)
@@ -244,7 +247,7 @@ function prescribed_perturbed_temperature_era5(
         method = time_interpolation_method,
         compose_function = compute_diffuse_fraction_broadcasted,
     )
-    perturb_lw_d(LW_d; ΔT = ΔT, a = 2) = LW_d .+ (a * ΔT)
+    perturb_lw_d(LW_d; ΔT = ΔT, a = 2) = max.(LW_d, Float32(0)) .+ (a * ΔT)
     LW_d = TimeVaryingInput(
         era5_ncdata_path,
         "msdwlwrf",
@@ -417,6 +420,9 @@ function prescribed_perturbed_rh_era5(
         start_date,
         regridder_type,
         regridder_kwargs = (; interpolation_method),
+        file_reader_kwargs = (;
+            preprocess_func = (data) -> max.(data, Float32(0)),
+        ),
         method = time_interpolation_method,
     )
     function compute_diffuse_fraction(total, direct)
@@ -444,6 +450,9 @@ function prescribed_perturbed_rh_era5(
         start_date,
         regridder_type,
         regridder_kwargs = (; interpolation_method),
+        file_reader_kwargs = (;
+            preprocess_func = (data) -> max.(data, Float32(0)),
+        ),
         method = time_interpolation_method,
     )
     cos_zenith_angle =