+++ src/standalone/Soil/spatially_varying_parameters.jl
@@ -19,7 +19,7 @@ masked_to_value(field, mask, value) = mask == 1.0 ? field : eltype(field)(value)
             Interpolations.Flat(),
         ),
         interpolation_method = Interpolations.Constant(),
-        lowres=ClimaLand.Domains.use_lowres_clm(surface_space),
+        lowres=false,
     )
 
 Reads spatially varying albedo parameters for the soil model, from NetCDF files
@@ -32,8 +32,7 @@ affect the regridding by (1) changing how we interpolate to ClimaCore points whi
 are not in the data, and (2) changing how extrapolate to points beyond the range of the
 data, and (3) changed the spatial interpolation method.
 The keyword argument lowres is a flag that determines if the 0.9x1.25 or 0.125x0.125
-resolution CLM data artifact is used. If the lowres flag is not provided, the clm artifact
-with the closest resolution to the surface_space is used.
+resolution CLM data artifact is used.
 
 Since these parameters are read from discretized data sets,
 they carry an inherent land/sea mask. This land/sea mask may not match the
@@ -48,7 +47,7 @@ function clm_soil_albedo_parameters(
         Interpolations.Flat(),
     ),
     interpolation_method = Interpolations.Constant(),
-    lowres = ClimaLand.Domains.use_lowres_clm(surface_space),
+    lowres = false,
 )
     context = ClimaComms.context(surface_space)
     PAR_albedo_dry, NIR_albedo_dry, PAR_albedo_wet, NIR_albedo_wet = map(